Move native types to a native types file
