+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/player/types.native.ts
@@ -1,0 +1,12 @@
+/**
+ * WordPress dependencies
+ */
+import { BlockInstance } from '@wordpress/blocks';
+
+export type NativePlayerProps = {
+	html: string;
+	isRequestingEmbedPreview: boolean;
+	isSelected: boolean;
+	clientId: string;
+	insertBlocksAfter: ( blocks: BlockInstance[] ) => void;
+};
