+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/player/types.ts
@@ -1,8 +1,4 @@
 /**
- * WordPress dependencies
- */
-import { BlockInstance } from '@wordpress/blocks';
-/**
  * External dependencies
  */
 import { VideoBlockAttributes, VideoPreviewProps } from '../../types';
@@ -15,12 +11,4 @@ export type PlayerProps = {
 	setAttributes: ( attributes: VideoBlockAttributes ) => void;
 	preview: VideoPreviewProps;
 	isRequestingEmbedPreview: boolean;
-};
-
-export type NativePlayerProps = {
-	html: string;
-	isRequestingEmbedPreview: boolean;
-	isSelected: boolean;
-	clientId: string;
-	insertBlocksAfter: ( blocks: BlockInstance[] ) => void;
 };
